services/analysis-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-genai>=1.0.0
pydantic==2.5.3
python-dotenv==1.0.0
asyncpg==0.30.0
